Inlined RestClient
